data/requirements.txt
jikanpy-v4
requests
python-dotenv
flask
jsonify
